+++ CleanArchitecture/Tests/MockPresenterProtocol.swift
@@ -5,7 +5,6 @@ class MockPresenterProtocol:PresenterProtocol {
     var onDidLoad:(() -> Void)?
     var onWillAppear:(() -> Void)?
     var onDidAppear:(() -> Void)?
-    var onOrientationChanged:(() -> Void)?
     var interactor:MockInteractorProtocol!
     var viewModel:ViewModel!
     
@@ -21,9 +20,5 @@ class MockPresenterProtocol:PresenterProtocol {
     
     func didAppear() {
         self.onDidAppear?()
-    }
-    
-    func orientationChanged() {
-        self.onOrientationChanged?()
     }
 }