+++ Demo/Source/Application/Application.swift
@@ -7,8 +7,16 @@ import UIKit
         let window:UIWindow = UIWindow(frame:UIScreen.main.bounds)
         window.backgroundColor = UIColor.white
         window.makeKeyAndVisible()
-        window.rootViewController = Router()
+        window.rootViewController = self.makeRouter()
         self.window = window
         return true
+    }
+    
+    private func makeRouter() -> UINavigationController {
+        let router:UINavigationController = UINavigationController()
+        let view:DemoView = DemoView()
+        view.presenter.interactor.router = router
+        router.setViewControllers([view], animated:false)
+        return router
     }
 }