+++ Demo/Source/Application/Presenter.swift
@@ -9,21 +9,21 @@ class Presenter:PresenterProtocol {
     
     func updatedRate(string:String) {
         let exchange:ExchangeText = self.interactor.updateRateWith(string:string)
-        var property:ViewModelUsd = ViewModelUsd()
+        var property:UsdViewModel = UsdViewModel()
         property.amount = exchange.usd
         self.viewModel.update(property:property)
     }
     
     func updatedEuro(string:String) {
         let exchange:ExchangeText = self.interactor.updateEuroWith(string:string)
-        var property:ViewModelUsd = ViewModelUsd()
+        var property:UsdViewModel = UsdViewModel()
         property.amount = exchange.usd
         self.viewModel.update(property:property)
     }
     
     func updatedUsd(string:String) {
         let exchange:ExchangeText = self.interactor.updateUsdWith(string:string)
-        var property:ViewModelEuro = ViewModelEuro()
+        var property:EuroViewModel = EuroViewModel()
         property.amount = exchange.euro
         self.viewModel.update(property:property)
     }
@@ -46,22 +46,28 @@ class Presenter:PresenterProtocol {
     
     private func initialRate() {
         let exchange:ExchangeText = self.interactor.initialExchanger()
-        var property:ViewModelRate = ViewModelRate()
+        var property:RateViewModel = RateViewModel()
         property.amount = exchange.rate
         self.viewModel.update(property:property)
     }
     
     private func initialEuro() {
         let exchange:ExchangeText = self.interactor.initialExchanger()
-        var property:ViewModelEuro = ViewModelEuro()
+        var property:EuroViewModel = EuroViewModel()
         property.amount = exchange.euro
         self.viewModel.update(property:property)
     }
     
     private func initialUsd() {
         let exchange:ExchangeText = self.interactor.initialExchanger()
-        var property:ViewModelUsd = ViewModelUsd()
+        var property:UsdViewModel = UsdViewModel()
         property.amount = exchange.usd
         self.viewModel.update(property:property)
     }
 }
+
+private struct Constants {
+    static let initialRate:Float = 1.17
+    static let initialEuro:Float = 1
+}
+