+++ Demo/Source/Domain Specific/Amount.swift
@@ -27,3 +27,10 @@ class Amount {
         return text
     }
 }
+
+private struct Constants {
+    static let defaultAmount:Float = 0
+    static let minIntegers:Int = 1
+    static let minDecimals:Int = 2
+    static let maxDecimals:Int = 2
+}