+++ Demo/Tests/TestView_ViewModel.swift
@@ -2,35 +2,50 @@ import XCTest
 @testable import Demo
 
 class TestView_ViewModel:XCTestCase {
-    private var view:View!
+    private var view:DemoView!
     
     override func setUp() {
         super.setUp()
-        self.view = View()
+        self.view = DemoView()
         self.view.viewDidLoad()
     }
     
     func testUpdateRateWithViewModel() {
-        self.view.content.inputRate.field.text = String()
-        var property:ViewModelRate = ViewModelRate()
+        let expect:XCTestExpectation = self.expectation(description:"Observer not updated")
+        self.view.content.rate.field.text = String()
+        var property:RateViewModel = RateViewModel()
         property.amount = "lorem ipsum"
         self.view.presenter.viewModel.update(property:property)
-        XCTAssertEqual(self.view.content.inputRate.field.text, property.amount, "Not updated")
+        DispatchQueue.main.async {
+            XCTAssertEqual(self.view.content.rate.field.text, property.amount, "Not updated")
+            expect.fulfill()
+        }
+        self.waitForExpectations(timeout:0.3, handler:nil)
     }
     
     func testUpdateEuroWithViewModel() {
-        self.view.content.inputEuro.field.text = String()
-        var property:ViewModelEuro = ViewModelEuro()
+        let expect:XCTestExpectation = self.expectation(description:"Observer not updated")
+        self.view.content.euro.field.text = String()
+        var property:EuroViewModel = EuroViewModel()
         property.amount = "lorem ipsum"
         self.view.presenter.viewModel.update(property:property)
-        XCTAssertEqual(self.view.content.inputEuro.field.text, property.amount, "Not updated")
+        DispatchQueue.main.async {
+            XCTAssertEqual(self.view.content.euro.field.text, property.amount, "Not updated")
+            expect.fulfill()
+        }
+        self.waitForExpectations(timeout:0.3, handler:nil)
     }
     
     func testUpdateUsdWithViewModel() {
-        self.view.content.inputUsd.field.text = String()
-        var property:ViewModelUsd = ViewModelUsd()
+        let expect:XCTestExpectation = self.expectation(description:"Observer not updated")
+        self.view.content.usd.field.text = String()
+        var property:UsdViewModel = UsdViewModel()
         property.amount = "lorem ipsum"
         self.view.presenter.viewModel.update(property:property)
-        XCTAssertEqual(self.view.content.inputUsd.field.text, property.amount, "Not updated")
+        DispatchQueue.main.async {
+            XCTAssertEqual(self.view.content.usd.field.text, property.amount, "Not updated")
+            expect.fulfill()
+        }
+        self.waitForExpectations(timeout:0.3, handler:nil)
     }
 }